fixed some file cleanup issues and defensive coding.

+++ bloxx/job-bloxx-update.sh
@@ -3,13 +3,18 @@
 ../bin/ETF-facts-update-common-job.sh \
     --collectionScript ./collectionScript-facts.sh --nodeArg bloxx-ETFs.csv "$@"
 
-cat links.json |
-    jq -r '.[] | [.ticker,.url] | @csv' | tr -d '"' | # tee tickerAndUrls.csv |
-    while IFS= read -r fundReference; do
-        ticker=$(echo "$fundReference" | cut -d, -f1 | tr -d '"')
-        echo "$fundReference" |
-            ../bin/ETF-distro-update-common-job.sh --ticker "$ticker" "$@"
-    done
+linksFile=links.json    
+if [ -f $linksFile ]; then
+    cat $linksFile |
+        jq -r '.[] | [.ticker,.url] | @csv' | tr -d '"' | # tee tickerAndUrls.csv |
+        while IFS= read -r fundReference; do
+            ticker=$(echo "$fundReference" | cut -d, -f1 | tr -d '"')
+            echo "$fundReference" |
+                ../bin/ETF-distro-update-common-job.sh --ticker "$ticker" "$@"
+        done
+else
+    echo $linksFile: missing!
+fi
 
 factsFile="history/bloxx-facts-new.json"
 if [ -f "$factsFile" ]; then
@@ -18,5 +23,7 @@ if [ -f "$factsFile" ]; then
             --accountClass Funds \
             --processScript ../lib/node-filter-yield-attrs.js \
             -f "$@"
+else
+    echo $factsFile: missing
 fi
-rm -f links.json
+rm -f $linksFile
